+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added auto-increment for episode number. It considers enabled/disabled items, trashed items, saved drafts and if episodeNumber is translatable, when searches for current maximum episode number.
 - Podcast taxonomies can now be bulk imported for german language.
 
 ## 0.17.0 - 2023-08-18